+++ lib/smart_pagination/widgets/smart_pagination_column.dart
@@ -59,6 +59,7 @@ class SmartPaginationColumn<T, R extends PaginationRequest> extends SmartPaginat
     super.insertItemAnimationBuilder,
     super.removeItemAnimationBuilder,
     super.animationDuration,
+    super.preserveScrollAnchorOnAppend,
     super.refreshListener,
     super.filterListeners,
     super.onInsertionCallback,
@@ -108,6 +109,7 @@ class SmartPaginationColumn<T, R extends PaginationRequest> extends SmartPaginat
     super.insertItemAnimationBuilder,
     super.removeItemAnimationBuilder,
     super.animationDuration,
+    super.preserveScrollAnchorOnAppend,
     super.refreshListener,
     super.filterListeners,
   }) : super.columnWithCubit();