+++ lib/smart_pagination/widgets/smart_pagination_grid_view.dart
@@ -64,6 +64,7 @@ class SmartPaginationGridView<T, R extends PaginationRequest> extends SmartPagin
     super.insertItemAnimationBuilder,
     super.removeItemAnimationBuilder,
     super.animationDuration,
+    super.preserveScrollAnchorOnAppend,
     super.refreshListener,
     super.filterListeners,
     super.onInsertionCallback,
@@ -119,6 +120,7 @@ class SmartPaginationGridView<T, R extends PaginationRequest> extends SmartPagin
     super.insertItemAnimationBuilder,
     super.removeItemAnimationBuilder,
     super.animationDuration,
+    super.preserveScrollAnchorOnAppend,
     super.refreshListener,
     super.filterListeners,
   }) : super.gridViewWithCubit();