+++ lib/smart_pagination/widgets/smart_pagination_list_view.dart
@@ -61,6 +61,7 @@ class SmartPaginationListView<T, R extends PaginationRequest> extends SmartPagin
     super.insertItemAnimationBuilder,
     super.removeItemAnimationBuilder,
     super.animationDuration,
+    super.preserveScrollAnchorOnAppend,
     super.refreshListener,
     super.filterListeners,
     super.onInsertionCallback,
@@ -115,6 +116,7 @@ class SmartPaginationListView<T, R extends PaginationRequest> extends SmartPagin
     super.insertItemAnimationBuilder,
     super.removeItemAnimationBuilder,
     super.animationDuration,
+    super.preserveScrollAnchorOnAppend,
     super.refreshListener,
     super.filterListeners,
   }) : super.listViewWithCubit();