+++ lib/smart_pagination/widgets/smart_pagination_page_view.dart
@@ -62,6 +62,7 @@ class SmartPaginationPageView<T, R extends PaginationRequest> extends SmartPagin
     super.insertItemAnimationBuilder,
     super.removeItemAnimationBuilder,
     super.animationDuration,
+    super.preserveScrollAnchorOnAppend,
     super.refreshListener,
     super.filterListeners,
     super.onInsertionCallback,
@@ -117,6 +118,7 @@ class SmartPaginationPageView<T, R extends PaginationRequest> extends SmartPagin
     super.insertItemAnimationBuilder,
     super.removeItemAnimationBuilder,
     super.animationDuration,
+    super.preserveScrollAnchorOnAppend,
     super.refreshListener,
     super.filterListeners,
   }) : super.pageViewWithCubit();